Restructuring string_from_apk and adding functionality to find activity by parsing AndroidManifest.xml, also including latest apk_tool
