+++ build.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 mvn clean compile assembly:single
 
-cp ./target/strings_from_apk-0.0.1-SNAPSHOT-jar-with-all.jar ./strings_from_apk.jar
+cp ./target/appium_apk_tools-0.0.2-SNAPSHOT-jar-with-all.jar ./appium_apk_tools.jar
