+++ pom.xml
@@ -1,10 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>strings_from_apk</groupId>
-	<artifactId>strings_from_apk</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>strings_from_apk</name>
+	<groupId>appium_apk_tools</groupId>
+	<artifactId>appium_apk_tools</artifactId>
+	<version>0.0.2-SNAPSHOT</version>
+	<name>appium_apk_tools</name>
 	<!-- mvn clean compile assembly:single -->
 	<packaging>jar</packaging>
 	<description>Extract strings.xml from an APK and convert to JSON</description>
@@ -34,7 +34,7 @@
 				<configuration>
 					<archive>
 						<manifest>
-							<mainClass>strings.StringsXML</mainClass>
+							<mainClass>io.appium.apktools.Main</mainClass>
 						</manifest>
 					</archive>
 					<descriptors>
